backend/requirements.txt
absl-py==2.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
astunparse==1.6.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.121.2
fastapi-cli==0.0.16
fastapi-cloud-cli==0.3.1
flatbuffers==25.9.23
gast==0.6.0
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.12.0
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
numpy==2.2.6
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.4
packaging==25.0
pillow==12.0.0
protobuf==6.33.1
pydantic==2.12.4
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.15.4
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.6
sentry-sdk==2.44.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
tomli==2.3.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==2.0.1
